+++ index.html
@@ -6,7 +6,6 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script src="js/functions.js"></script>
   <title>Кекстаграм</title>
 </head>
 
@@ -227,6 +226,6 @@
       </div>
     </section>
   </template>
-
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
